backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML Providers
google-genai>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.26.0

# Image Processing
Pillow>=10.2.0
rembg[cpu]>=2.0.50

# Data & Storage
pydantic>=2.6.0
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
simpleeval>=1.0.0       # Safe expression evaluation for pipeline conditions
rich>=13.0.0            # Beautiful CLI output
click>=8.1.0            # CLI framework

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
litellm>=1.0.0
